+++ services/py/install.sh
@@ -1,0 +1,10 @@
+./pip install savReaderWriter
+./pip install ijson
+./pip install numpy
+
+# En hosting compartido hacer antes:
+#> virtualenv python
+#> cd ~/mapas/bin y... ejecutar ~/mapas/bin/pip install savReaderWriter
+
+# para ver la carpeta actual de python:
+# > which python