+++ services/py/spss2json.py
@@ -1,10 +1,14 @@
 # -*- coding: utf-8 -*-
 
+import os
+os.environ["MKL_NUM_THREADS"] = "1"
+os.environ["NUMEXPR_NUM_THREADS"] = "1"
+os.environ["OMP_NUM_THREADS"] = "1"
+
 import savReaderWriter
 import traceback
 import json
 import sys
-import os
 import re
 
 def main():