+++ services/src/services/backoffice/publish/snapshots/Variable.php
@@ -146,7 +146,7 @@ class Variable
 			throw new ErrorException("La variable '" . $this->attributes['mvv_caption']. "' de la métrica "
 				. $this->GetVariableMetricErrorCaption() . " no tiene valores. Revise la symbología de la variable.");
 		//
-		if (sizeof($values) == 1 && sizeof($values[0]['vvl_value']) == null)
+		if (sizeof($values) == 1 && $values[0]['vvl_value'] == null)
 			return $values[0]['vvl_id'];
 
 		$cutMode = $this->attributes['vsy_cut_mode'];